+++ changelog.md
@@ -944,5 +944,6 @@ Security in case of vulnerabilities.
   - detect less and scss modification on all files
   - elFinder - Version 2.1.61 
   - template editor, save and commit with ajax request
+  - exrtact anchor id from heading
 ### fixed
   - link without scheme usable in field external link